change group number
