+++ src/fa35/group2/Main.java
@@ -1,4 +1,4 @@
-package fa35.group1;
+package fa35.group2;
 
 public class Main {
     public static void main(String[] args) {
