+++ src/fa35/group2/model/SkillEntity.java
@@ -1,12 +1,13 @@
-package fa35.group1.model;
+package fa35.group2.model;
 
 import javax.persistence.*;
 
-@Entity(name = "payment")
-@NamedQuery(name = "payment.get", query = "SELECT c FROM payment c")
-public class PaymentEntity {
+@Entity(name = "skill")
+@NamedQuery(name = "skill.get", query = "SELECT c FROM skill c")
+public class SkillEntity {
+
     @Id
-    @Column(name = "payment_id")
+    @Column(name = "skill_id")
     @GeneratedValue
     private int id;
 
@@ -30,9 +31,8 @@ public class PaymentEntity {
 
     @Override
     public boolean equals(Object obj) {
-        if (obj instanceof PaymentEntity) {
-            PaymentEntity other = (PaymentEntity) obj;
-
+        if (obj instanceof SkillEntity) {
+            SkillEntity other = (SkillEntity) obj;
             return (this.getId() == other.getId()) &&
                     ((this.getName() == null && other.getName() == null) ||
                             (this.getName() != null && this.getName().equals(other.getName())));
