+++ src/fa35/group2/model/xml/XmlPersistence.java
@@ -1,9 +1,9 @@
-package fa35.group1.model.xml;
+package fa35.group2.model.xml;
 
-import fa35.group1.model.FriendEntity;
-import fa35.group1.model.IPersistence;
-import fa35.group1.model.PaymentEntity;
-import fa35.group1.model.SkillEntity;
+import fa35.group2.model.FriendEntity;
+import fa35.group2.model.IPersistence;
+import fa35.group2.model.PaymentEntity;
+import fa35.group2.model.SkillEntity;
 
 import java.util.List;
